better documentatin and some renaming
