+++ lammy-core/src/main/java/com/sigwned/lammy/core/serialization/PlatformCustomPojoSerializer.java
@@ -31,9 +31,14 @@ import com.amazonaws.services.lambda.runtime.serialization.PojoSerializer;
 import com.amazonaws.services.lambda.runtime.serialization.events.LambdaEventSerializers;
 import com.amazonaws.services.lambda.runtime.serialization.factories.GsonFactory;
 import com.amazonaws.services.lambda.runtime.serialization.factories.JacksonFactory;
+import com.sigwned.lammy.core.base.bean.BeanLambdaFunctionBase;
+import com.sigwned.lammy.core.base.streamedbean.StreamedBeanLambdaFunctionBase;
 
 /**
- * The default {@link PojoSerializer serializer} used by the AWS Lambda platform.
+ * The default {@link PojoSerializer serializer} used by the AWS Lambda platform. Note that it is
+ * {@link ContextAwareCustomPojoSerializer context-aware}, which means it can only be used in
+ * application-serialized Lambda functions (e.g., {@link StreamedBeanLambdaFunctionBase}) and not in
+ * platform-serialized Lambda functions (e.g., {@link BeanLambdaFunctionBase}).
  */
 public class PlatformCustomPojoSerializer implements ContextAwareCustomPojoSerializer {
   private static enum Platform {
